Remove unused measure controller
