+++ src/components/measure/MeasureModule.js
@@ -4,7 +4,6 @@ goog.require('ga_measure_service');
 (function() {
 
   angular.module('ga_measure', [
-    'ga_measure_service',
     'ga_measure_directive',
     'ga_measure_service'
   ]);
