+++ src/js/DrawController.js
@@ -1,13 +1,20 @@
 goog.provide('ga_draw_controller');
+
+goog.require('ga_browsersniffer_service');
+goog.require('ga_styles_service');
+goog.require('ga_print_service');
 (function() {
 
   var module = angular.module('ga_draw_controller', [
     'pascalprecht.translate',
-    'ga_styles_service'
+    'ga_styles_service',
+    'ga_browsersniffer_service',
+    'ga_print_service'
   ]);
 
   module.controller('GaDrawController',
-      function($rootScope, $scope, $translate, gaGlobalOptions, gaStyleFactory) {
+      function($rootScope, $scope, $translate, $timeout, gaBrowserSniffer,
+      gaGlobalOptions, gaStyleFactory, gaPrintService) {
         
         $scope.$on('gaPopupFocusChange', function(evt, isFocus) {
           $scope.options.hasPopupFocus = isFocus;
@@ -486,5 +493,30 @@ goog.provide('ga_draw_controller');
           tool.activeKey = 'is' + tool.id.charAt(0).toUpperCase() + tool.id.slice(1) + 'Active';
           tool.title = 'draw_' + tool.id;
         }
+
+        // Allow to print dynamic profile from feature's popup
+        // TODO: Verify f it's working, currently print profile is deactivated.
+        $scope.print = function() {
+          var contentEl = $('ga-draw-popup .ga-popup-content');
+          var onLoad = function(printWindow) {
+            var profile = $(printWindow.document).find('[ga-profile]');
+            // HACK IE, for some obscure reason an A4 page in IE is not
+            // 600 pixels width so calculation of the scale is not optimal.
+            var b = (gaBrowserSniffer.msie) ? 1000 : 600;
+            // Same IE mistery here, a js error occurs using jQuery width() function.
+            var a = parseInt(profile.find('svg').attr('width'), 10);
+            var scale = b / a;
+            profile.css({
+              position: 'absolute',
+              left: (-(a - a * scale) / 2) + 'px',
+              top: '200px',
+              transform: 'scale(' + scale + ')'
+            });
+            printWindow.print();
+          }
+          $timeout(function() {
+            gaPrintService.htmlPrintout(contentEl.clone().html(), undefined, onLoad);
+          }, 0, false);
+        };
       });
 })();
