+++ src/js/GaModule.js
@@ -55,7 +55,6 @@ goog.require('ga_seo_controller');
 goog.require('ga_timeselector_controller');
 goog.require('ga_tooltip_controller');
 goog.require('ga_featuretree_controller');
-goog.require('ga_measure_controller');
 goog.require('ga_draw_controller');
 goog.require('ga_query_controller');
 (function() {
@@ -116,7 +115,6 @@ goog.require('ga_query_controller');
     'ga_timeselector_controller',
     'ga_tooltip_controller',
     'ga_featuretree_controller',
-    'ga_measure_controller',
     'ga_draw_controller',
     'ga_query_controller'
   ]);
